+++ example/a.ts
@@ -1,32 +1,18 @@
-import {
-	sucrose,
-	separateFunction,
-	findParameterReference,
-	extractMainParameter,
-	isContextPassToFunction
-} from '../src/sucrose'
+// index.ts
+import { Elysia, t } from '../src'
 
-const event = `async({user:z,params:{id:J},query:{conversation:G}})=>{const Q=await z.id;if(G)return setImmediate(()=>{O6.conversation.setActiveConversation(Q,J,G)}),O6.conversation.getChatsById(Q,J,G);return O6.conversation.getChats(Q,J)}`
+const app = new Elysia()
+	.guard({
+		cookie: t.Cookie({ session: t.String() })
+	})
+	.get('/', ({ cookie: { session } }) =>
+		session.value ? session.value : 'Empty'
+	)
 
-const [parameter, body, { isArrowReturn }] = separateFunction(event.toString())
+let response = await app.handle(new Request(`http://localhost`))
+console.log(await response.text())
 
-const inference = {
-	body: false,
-	cookie: false,
-	headers: false,
-	query: false,
-	server: false,
-	set: false
-}
-
-const rootParameters = findParameterReference(parameter, {
-	body: false,
-	cookie: false,
-	headers: false,
-	query: false,
-	server: false,
-	set: false
-})
-const mainParameter = extractMainParameter(rootParameters)
-
-isContextPassToFunction(mainParameter!, body, inference)
+response = await app.handle(
+	new Request(`http://localhost`, { headers: { Cookie: 'session=value' } })
+)
+console.log(await response.text())